+++ clang/include/clang/Parse/ParseHLSLRootSignature.h
@@ -17,6 +17,7 @@
 #include "clang/Basic/DiagnosticParse.h"
 #include "clang/Lex/LexHLSLRootSignature.h"
 #include "clang/Lex/Preprocessor.h"
+#include "clang/Sema/SemaHLSL.h"
 
 #include "llvm/ADT/SmallVector.h"
 #include "llvm/ADT/StringRef.h"
@@ -25,21 +26,6 @@
 
 namespace clang {
 namespace hlsl {
-
-// Introduce a wrapper struct around the underlying RootElement. This structure
-// will retain extra clang diagnostic information that is not available in llvm.
-struct RootSignatureElement {
-  RootSignatureElement(SourceLocation Loc,
-                       llvm::hlsl::rootsig::RootElement Element)
-      : Loc(Loc), Element(Element) {}
-
-  const llvm::hlsl::rootsig::RootElement &getElement() const { return Element; }
-  const SourceLocation &getLocation() const { return Loc; }
-
-private:
-  SourceLocation Loc;
-  llvm::hlsl::rootsig::RootElement Element;
-};
 
 class RootSignatureParser {
 public:
